requirements.txt
pandas
numpy
h3
shapely
pyproj
python-dotenv
supabase
requests
